refactor(ui): resolve a device's install key behind one hook
DeviceDetails paired useInstallKeys with resolveEnrollmentSource and repeated the page size
the lookup depends on. A second caller was about to repeat both, so the pair moves behind
useEnrollmentSource and the page size stops being a literal each caller has to agree on.

The lookup is client-side over a single page of keys, so a namespace past that page resolves
its tail to null rather than to a name. That limit is why the hook's doc comment states it:
it is not visible from the call site, and a caller that assumes otherwise renders a dash.

KeyValueChip moves to components/common for the same reason — a second page tree needs it,
and reaching across into pages/install-keys/ to get it is not a dependency worth having.
