+++ ui/apps/console/src/components/common/KeyValueChip.tsx
